Add term to model to account for inverter power limits
Over the summer, the panels can generate much more power than the 5 kW
limit of the inverter. This causes a flattening of the energy generation
curve compared to the previous sinusoidal model. Here, we introduce a
saturating `tanh()` term to account for this, effectively
`tanh(cos(t))`.
  This is of course a major oversimplification, though a more plausible
model would most likely be substantially more complex. Loosely, there
are two components to optimal energy generation, namely solar irradiance
and length of day (with total daily production relating to the former
integrated over the latter). However, the inverter saturating only
affects the irradiance. In other words, the total energy (i.e. the area
under the power curve) would keep increasing proportionally with the
length of the day but only sublinearly with irradiance once the inverter
is maxed out.