Relax prior over weather effect
This tweaks the prior distribution such that the probability tends to
zero as the weather effect approaches one. This not only makes the
sampler happier without the hard transition to the region of invalid
samples, but also is more realistic. This does have the effect of
increasing the uncertainty of the amplitude parameter and pushing the
maximum further from the observed values but neither of these are
necessarily problematic.