Reparametrise the limit parameter for inverter clipping
If the limit is sampled such that it is less than any of the data points
then the model will not be invertible and the samples will have to be
dropped. Suspicion is that this is causing issues for convergence of
these parameters.
  New scheme sets a minimum value for this parameter that just needs to
be set to be greater than the data max. While this is not terribly
principled from a Bayesian standpoint, it does improve model fit for the
summer months when production is clipped.