Change weather effect prior to mixture of gammas
This, coupled with the change to the limit in the previous commit, all
but eliminates the numerical errors from the sampler. The gamma allows
for weather effect parameters which are slightly above 1.0 (e.g. if the
parameters for `E_{avail}` are slightly low). While having a hard upper
limit is somewhat appealing for the interpretation of the max production
values, this had already been relaxed by having the beta distributions
no longer butting up against 1.0. Similarly, it is possible to exceed
the 'cloudless day' production as e.g. clouds not obscuring the sun will
scatter extra panels onto the panels.