Add data from May 2024
Have had to tighten up the priors as was getting a second mode of the
posterior, with higher log probabilities, but at a physically
implausible setting of the saturation function. The problem is that
`log_sum_exp()` as we use it gives us a smooth interpolation between the
linear and saturated regime, but the higher the smoothness parameter the
further the saturated values lie below the true piecewise linear
function we are interpolating. At very high smoothness values this can
cause the saturated production to go negative for very small inputs,
which is clearly physically implausible. However, the inverse of this
artificially inflates the weather effect parameters when the production
is very low, and, as low values are less likely a priori, this makes
this physically implausible setting more probable a posteriori. At some
stage this will require a tweak to the saturation function to prevent
issues like this.