Rework saturation function
As noted in `d7a3a78`, there were some conceptual issues with the
log-sum-exp based saturation function, namely that for large values of
the smoothness parameter the saturated output could be less than zero
for small inputs. We require a function that starts out as f(x) ≈ x
before dropping off. The old tanh function had this property, but the
smoothness could not be controlled. We introduce a new form of this
here, where we essentially have f(x) = tanh(sinh(x)). Compared to the
standard tanh, the internal sinh pushes larger values of x further into
the tails of the tanh function, thereby causing it to saturate faster
and it is possible to add a parameter to control this.
  As this is more stable we can also revert some of the changes to the
priors that had been used to bias against the physically implausible
regime of the log-sum-exp function.